pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "medmatch"
version = "0.1.0"
description = "MedMatch experiment drivers and shared evaluation utilities."
requires-python = ">=3.9"
dependencies = [
  "openpyxl",
  "google-genai",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
